backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
Pillow>=10.4.0
whitenoise==6.6.0
dj-database-url==2.2.0
gunicorn==23.0.0
